# docs/swagger.yaml
basePath: /
definitions:
  models.Category:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        description: Код ошибки, например, 400 или 500
        type: integer
      message:
        description: Сообщение об ошибке
        type: string
    type: object
  models.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  models.Product:
    properties:
      category_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      manufacturer:
        type: string
      name:
        type: string
      price:
        type: number
      rating:
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: Support Team
    url: http://example.com/support
  description: API для интернет-магазина спортивного питания. Содержит функционал
    для управления пользователями, продуктами, заказами и отзывами.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: Sports Nutrition Store API
  version: "1.0"
paths:
...
  /categories:
    get:
      consumes:
      - application/json
      description: Возвращает список категорий с предзагрузкой связанных продуктов,
        ограничивая время выполнения запроса до 2 секунд.
      parameters:
      - description: токен
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список категорий с предзагруженными продуктами
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
        "408":
          description: Тайм-аут запроса
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение списка категорий с тайм-аутом
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Создает новую категорию в базе данных на основе переданных данных.
      parameters:
      - description: токен
        in: header
        name: Authorization
        type: string
      - description: Данные категории
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Созданная категория
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создание новой категории
      tags:
      - categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет категорию по переданному ID
      parameters:
      - description: токен
        in: header
        name: Authorization
        type: string
      - description: Идентификатор категории
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Категория успешно удалена
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "404":
          description: Категория не найдена
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление категории
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: Возвращает информацию о категории на основе переданного идентификатора.
      parameters:
      - description: токен
        in: header
        name: Authorization
        type: string
      - description: Идентификатор категории
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Информация о категории
          schema:
            $ref: '#/definitions/models.Category'
        "404":
          description: Категория не найдена
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение категории по ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Обновляет категорию с переданными данными на основе ID
      parameters:
      - description: токен
        in: header
        name: Authorization
        type: string
      - description: Идентификатор категории
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные категории
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "200":
          description: Категория успешно обновлена
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Категория не найдена
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновление категории
      tags:
      - categories
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...